Shift Desanitizer methods to Sanitizer
