+++ src/main/java/teammates/storage/search/CommentSearchDocument.java
@@ -9,11 +9,11 @@ import teammates.common.datatransfer.CourseAttributes;
 import teammates.common.datatransfer.InstructorAttributes;
 import teammates.common.datatransfer.StudentAttributes;
 import teammates.common.util.Const;
-import teammates.common.util.Desanitizer;
 import teammates.common.util.JsonUtils;
 
 import com.google.appengine.api.search.Document;
 import com.google.appengine.api.search.Field;
+import teammates.common.util.Sanitizer;
 
 /**
  * The SearchDocument object that defines how we store {@link Document} for student comments
@@ -63,7 +63,7 @@ public class CommentSearchDocument extends SearchDocument {
         case TEAM:
             for (String team : comment.recipients) {
                 List<StudentAttributes> students =
-                        logic.getStudentsForTeam(Desanitizer.desanitizeFromHtml(team), comment.courseId);
+                        logic.getStudentsForTeam(Sanitizer.desanitizeFromHtml(team), comment.courseId);
                 if (students != null) {
                     relatedStudents.addAll(students);
                 }
