+++ src/main/java/teammates/ui/automated/AdminSendEmailWorkerAction.java
@@ -4,8 +4,8 @@ import teammates.common.datatransfer.AdminEmailAttributes;
 import teammates.common.exception.TeammatesException;
 import teammates.common.util.Assumption;
 import teammates.common.util.Const.ParamsNames;
-import teammates.common.util.Desanitizer;
 import teammates.common.util.EmailWrapper;
+import teammates.common.util.Sanitizer;
 import teammates.logic.api.EmailGenerator;
 
 /**
@@ -48,7 +48,7 @@ public class AdminSendEmailWorkerAction extends AutomatedAction {
         
         try {
             EmailWrapper email =
-                    new EmailGenerator().generateAdminEmail(Desanitizer.desanitizeFromHtml(emailContent),
+                    new EmailGenerator().generateAdminEmail(Sanitizer.desanitizeFromHtml(emailContent),
                                                             emailSubject, receiverEmail);
             emailSender.sendEmail(email);
             log.info("Email sent to " + receiverEmail);
