+++ src/main/java/teammates/ui/pagedata/FeedbackSubmissionEditPageData.java
@@ -12,7 +12,7 @@ import teammates.common.datatransfer.InstructorAttributes;
 import teammates.common.datatransfer.StudentAttributes;
 import teammates.common.util.Config;
 import teammates.common.util.Const;
-import teammates.common.util.Desanitizer;
+import teammates.common.util.Sanitizer;
 import teammates.common.util.StringHelper;
 import teammates.ui.template.FeedbackSubmissionEditQuestion;
 import teammates.ui.template.FeedbackSubmissionEditResponse;
@@ -180,7 +180,7 @@ public class FeedbackSubmissionEditPageData extends PageData {
                    + "</option>");
         
         for (Map.Entry<String, String> pair : emailNamePair.entrySet()) {
-            boolean isSelected = Desanitizer.desanitizeFromHtml(pair.getKey())
+            boolean isSelected = Sanitizer.desanitizeFromHtml(pair.getKey())
                                              .equals(currentlySelectedOption);
             result.add("<option value=\"" + sanitizeForHtml(pair.getKey()) + "\"" + (isSelected ? " selected" : "") + ">"
                            + sanitizeForHtml(pair.getValue())
