+++ src/main/java/teammates/ui/pagedata/InstructorCommentsPageData.java
@@ -10,7 +10,7 @@ import teammates.common.datatransfer.CommentAttributes;
 import teammates.common.datatransfer.CourseRoster;
 import teammates.common.datatransfer.FeedbackSessionAttributes;
 import teammates.common.datatransfer.InstructorAttributes;
-import teammates.common.util.Desanitizer;
+import teammates.common.util.Sanitizer;
 import teammates.ui.template.CommentRow;
 import teammates.ui.template.CommentsForStudentsTable;
 import teammates.ui.template.CoursePagination;
@@ -119,13 +119,13 @@ public class InstructorCommentsPageData extends PageData {
     private List<CommentRow> createCommentRows(
             String giverEmail, String giverName, List<CommentAttributes> commentsForGiver,
             Map<String, List<Boolean>> commentModifyPermissions, CourseRoster roster) {
-        String unsanitizedGiverName = Desanitizer.desanitizeFromHtml(giverName);
+        String unsanitizedGiverName = Sanitizer.desanitizeFromHtml(giverName);
         
         List<CommentRow> rows = new ArrayList<CommentRow>();
         for (int i = 0; i < commentsForGiver.size(); i++) {
             CommentAttributes comment = commentsForGiver.get(i);
             String recipientDetails = getRecipientNames(comment.recipients, courseId, null, roster);
-            String unsanitizedRecipientDetails = Desanitizer.desanitizeFromHtml(recipientDetails);
+            String unsanitizedRecipientDetails = Sanitizer.desanitizeFromHtml(recipientDetails);
             
             Boolean isInstructorAllowedToModifyCommentInSection = commentModifyPermissions.get(giverEmail).get(i);
             String typeOfPeopleCanViewComment = getTypeOfPeopleCanViewComment(comment);
